refactor: Restructure test config
