+++ tests/config.py
@@ -1,0 +1,3 @@
+from pathlib import Path
+
+GENERATED_PATH = Path(__file__).parent / "generated"
