+++ tests/conftest.py
@@ -1,8 +1,4 @@
+from tests.config import GENERATED_PATH
 from tests.fixtures import *  # noqa: F401, F403
 
-DEVELOPMENT_MODE: bool = False
-
-GENERATED_PATH = os.path.sep.join(["tests", "generated"])
-
-if not os.path.exists(GENERATED_PATH):
-    os.makedirs(GENERATED_PATH)
+GENERATED_PATH.mkdir(parents=True, exist_ok=True)
